Fix a signed comparison

+++ include/occa/dtype/dtype.hpp
@@ -369,7 +369,7 @@ namespace occa {
                     );
       dtype_t& type = entry.first->second;
 
-      for (int i = 0; i < fieldNames.size(); ++i) {
+      for (size_t i = 0; i < fieldNames.size(); ++i) {
         type.addField(fieldNames[i], fieldTypes[i]);
       }
